refactor(uzicli): split account verbs into client_account.go/fake_account.go (PRD #1017 M3)
Pure motion: move the account/self verbs (whoami, token, memory, version,
CLI-auth) and pollStatusField helper, plus their FakeClient twins, into
per-domain files. go doc -u -all byte-identical; no test edits.
